+++ src/services/content/Service/Content.Service/Models/ContentOptionFilter.cs
@@ -8,16 +8,16 @@ namespace Content.Service.Models
         /// <summary>
         /// Gets or sets the content id which is filtered.
         /// </summary>
-        public int? ContentId { get; set; }
+        public ulong? ContentId { get; set; }
 
         /// <summary>
         /// Gets or sets the page id which is filtered.
         /// </summary>
-        public int? PageId { get; set; }
+        public ulong? PageId { get; set; }
 
         /// <summary>
         /// Gets or sets the book id which is filtered.
         /// </summary>
-        public int? BookId { get; set; }
+        public ulong? BookId { get; set; }
     }
 }
