+++ src/services/content/Service/Content.Service/ViewModels/ContentOptionFilter.cs
@@ -8,16 +8,16 @@ namespace Content.Service.ViewModels
         /// <summary>
         /// Gets or sets the content id which is filtered.
         /// </summary>
-        public int? ContentId { get; set; }
+        public ulong? ContentId { get; set; }
 
         /// <summary>
         /// Gets or sets the scene id which is filtered.
         /// </summary>
-        public int? PageId { get; set; }
+        public ulong? PageId { get; set; }
 
         /// <summary>
         /// Gets or sets the book id which is filtered.
         /// </summary>
-        public int? BookId { get; set; }
+        public ulong? BookId { get; set; }
     }
 }
