+++ src/services/workspace/Service/Workspace.Service/Controllers/PageController.cs
@@ -74,7 +74,7 @@ namespace Workspace.Service.Controllers
         [SwaggerResponse(StatusCodes.Status404NotFound, "A page with the specified id was not found.", typeof(ProblemDetails))]
         public Task<IActionResult> DeleteAsync(
             [FromServices] DeletePageCommand command,
-            int pageId,
+            ulong pageId,
             CancellationToken cancellationToken) => command.ExecuteAsync(pageId, cancellationToken);
 
         /// <summary>
@@ -153,7 +153,7 @@ namespace Workspace.Service.Controllers
         [SwaggerResponse(StatusCodes.Status415UnsupportedMediaType, "The MIME type in the Content-Type HTTP header is unsupported.", typeof(ProblemDetails))]
         public Task<IActionResult> PutAsync(
             [FromServices] PutPageCommand command,
-            int pageId,
+            ulong pageId,
             [FromBody] SavePage page,
             CancellationToken cancellationToken) => command.ExecuteAsync(pageId, page, cancellationToken);
     }
