+++ src/services/workspace/Service/Workspace.Service/Models/Book.cs
@@ -10,7 +10,7 @@
         /// <summary>
         /// Gets or sets the book id.
         /// </summary>
-        public int BookId { get; set; }
+        public ulong BookId { get; set; }
 
         /// <summary>
         /// Gets or sets the name.
