+++ src/services/workspace/Service/Workspace.Service/ViewModels/Book.cs
@@ -10,7 +10,7 @@ namespace Workspace.Service.ViewModels
         /// <summary>
         /// Gets or sets the book id.
         /// </summary>
-        public int BookId { get; set; } = default!;
+        public ulong BookId { get; set; } = default!;
 
         /// <summary>
         /// Gets or sets the name.
