+++ src/services/workspace/Service/Workspace.Service/ViewModels/PageOptionFilter.cs
@@ -8,11 +8,11 @@ namespace Workspace.Service.ViewModels
         /// <summary>
         /// Gets or sets the page id which is filtered.
         /// </summary>
-        public int? PageId { get; set; }
+        public ulong? PageId { get; set; }
 
         /// <summary>
         /// Gets or sets the book id which is filtered.
         /// </summary>
-        public int? BookId { get; set; }
+        public ulong? BookId { get; set; }
     }
 }
